Move user activity widget
